Sedikit perbaikan di modul PNBP & kelola persentase

+++ database/migrations/2017_12_14_100012_create_percentage_table.php
@@ -16,7 +16,7 @@ class CreatePercentageTable extends Migration
         Schema::create('percentage', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name', 64);
-            $table->integer('value');
+            $table->bigInteger('value');
             $table->timestamps();
         });
     }
